Directory restructure to better play with modules
Much cleaner this way!
