+++ package.json
@@ -20,6 +20,7 @@
     "gulp-mocha": "^2.0.0",
     "gulp-sourcemaps": "^1.3.0",
     "gulp-uglify": "^1.0.2",
+    "gulp-wrap": "^0.8.0",
     "jshint-stylish": "^1.0.0",
     "lodash": "^2.4.1"
   },
