+++ src/core/state/base.js
