+++ src/core/tokens/multi.js
