+++ src/core/tokens/text.js
