+++ src/linkify-with-hashtag.js
@@ -1,5 +1,5 @@
 // NOTE: This file should only be used to build into a browser package
 // Linkify with basic hashtags support
-let linkify = require('./linkify');
+import linkify from './linkify';
 require('./plugins/hashtag')(linkify);
-module.exports = linkify;
+export default linkify;
