+++ test/spec/core/state/character.js
@@ -1,7 +1,7 @@
 /*jshint -W030 */
 var
-TEXT_TOKENS = require('../../../lib/tokens/text'),
-CharacterState = require('../../../lib/scanner/state');
+TEXT_TOKENS = require('../../../../lib/core/tokens/text'),
+CharacterState = require('../../../../lib/core/state/character');
 
 describe('CharacterState', function () {
 	var S_START, S_DOT, S_NUM;
