+++ test/spec/core/tokens/text.js
@@ -1,4 +1,4 @@
-var TEXT_TOKENS = require('../../../lib/tokens/text');
+var TEXT_TOKENS = require('../../../../lib/core/tokens/text');
 
 describe('TEXT_TOKENS', function () {
